unify file downloads
* split download from FileMixin to FileDownloadMixin
* move FileMixin (and FileDownloadMixin) from leasing.viewsets.utils to utils.viewsets.mixins
*  use FileResponse in place of HttpResponse in DownloadMixin.download(), which handles the mimetype checking (from file extension), adds Content-Length header, and handles file closure
* replace duplicated download() methods in viewsets with FileDownloadMixin for: AttachmentViewSet, and AreaSearchAttachmentViewset
