+++ leasing/viewsets/utils.py
@@ -1,13 +1,8 @@
 import json
-import os
 
-from django.conf import settings
-from django.core.files import File
 from django.db import IntegrityError, transaction
-from django.http import HttpResponse
 from django.utils.translation import gettext_lazy as _
 from rest_framework import parsers, status, viewsets
-from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.views import exception_handler
 
@@ -39,66 +34,6 @@ class MultiPartJsonParser(parsers.MultiPartParser):
             files[key] = result.files.get(key)
 
         return parsers.DataAndFiles(data, files)
-
-
-class FileMixin:
-    def create(self, request, *args, **kwargs):
-        """Use the Class.serializer_class after the creation for returning the saved data.
-        Instead of a different serializer used in 'create' action."""
-        if not self.serializer_class:
-            return super().create(request, *args, **kwargs)
-
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-
-        read_serializer = self.serializer_class(
-            serializer.instance, context=serializer.context
-        )
-
-        headers = self.get_success_headers(read_serializer.data)
-        return Response(
-            read_serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
-
-    def update(self, request, *args, **kwargs):
-        """Use the Class.serializer_class after update for returning the saved data.
-        Instead of a different serializer used in 'update' action."""
-        if not self.serializer_class:
-            return super().create(request, *args, **kwargs)
-
-        partial = kwargs.pop("partial", False)
-        instance = self.get_object()
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-
-        if getattr(instance, "_prefetched_objects_cache", None):
-            # If 'prefetch_related' has been applied to a queryset, we need to
-            # forcibly invalidate the prefetch cache on the instance.
-            instance._prefetched_objects_cache = {}
-
-        read_serializer = self.serializer_class(
-            serializer.instance, context=serializer.context
-        )
-
-        return Response(read_serializer.data)
-
-    @action(methods=["get"], detail=True)
-    def download(self, request, pk=None):
-        obj = self.get_object()
-
-        filename = "/".join([settings.MEDIA_ROOT, obj.file.name])
-        base_filename = os.path.basename(obj.file.name)
-
-        with open(filename, "rb") as fp:
-            # TODO: detect file MIME type
-            response = HttpResponse(File(fp), content_type="application/octet-stream")
-            response["Content-Disposition"] = 'attachment; filename="{}"'.format(
-                base_filename
-            )
-
-            return response
 
 
 def integrityerror_exception_handler(exc, context):
